# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "test swagger middleware",
    "description": "a test swagger file"
  },
  "host": "test.com",
  "tags": [{ "name": "test", "description": "a test tag" }],
  "schemes": ["http", "https"],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "text",
    "text/plain"
  ],
  "produces": ["application/json"],
  "security": [],
  "securityDefinitions": {},
  "definitions": {
    "Success": { "type": "object", "example": { "message": "Success" } },
    "Error400": { "type": "object", "example": { "message": "Bad Request" } },
    "Error500": {
      "type": "object",
      "example": { "message": "Internal Server Error" }
    }
  },
  "paths": {
    "/test/{path}": {
      "post": {
        "operationId": "test",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": { "type": "object", "example": { "message": "Success" } }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "type": "object",
              "example": { "message": "Bad Request" }
            }
          },
          "500": {
            "description": "server error",
            "schema": {
              "type": "object",
              "example": { "message": "Internal Server Error" }
            }
          }
        }
      }
    },
    "/test/{path}/test2/{path2}": {
      "post": {
        "operationId": "test",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "type": "integer",
            "required": true
          },
          {
            "name": "path2",
            "in": "path",
            "type": "boolean",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": { "type": "object", "example": { "message": "Success" } }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "type": "object",
              "example": { "message": "Bad Request" }
            }
          },
          "500": {
            "description": "server error",
            "schema": {
              "type": "object",
              "example": { "message": "Internal Server Error" }
            }
          }
        }
      }
    }
  }
}